requirements.txt
streamlit
supabase-py
bcrypt
pytest

"lkvDujFoXSyIC6UY"